minor changes wrt file locations and file names - jobCategories constants are moved to helpers dir. - api.dart is renamed to graphql_queries.dart. - saved_jobs.dart is since provider is not used anymore.
