+++ lib/helpers/graphql_queries.dart
