+++ lib/services/graphql_service.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:graphql_flutter/graphql_flutter.dart';
-import 'package:moovup_demo/helpers/api.dart';
+import 'package:moovup_demo/helpers/graphql_queries.dart';
 import 'package:moovup_demo/models/search_option_data.dart';
 import 'package:moovup_demo/services/service.dart';
 import 'package:http/http.dart' as http;
